Check if speed limit files exist

+++ swraid.functions.sh
@@ -8,6 +8,8 @@
 
 # suspend_swraid_resync
 suspend_swraid_resync() {
+  [[ -e /proc/sys/dev/raid/speed_limit_max || -e /proc/sys/dev/raid/speed_limit_min ]] || return
+
   echo 0 | tee /proc/sys/dev/raid/speed_limit_max > /proc/sys/dev/raid/speed_limit_min
 }
 
@@ -25,6 +27,8 @@ unset_raid0_default_layout() {
 
 # resume_swraid_resync
 resume_swraid_resync() {
+  [[ -e /proc/sys/dev/raid/speed_limit_max || -e /proc/sys/dev/raid/speed_limit_min ]] || return
+
   echo 200000 > /proc/sys/dev/raid/speed_limit_max
   echo 1000 > /proc/sys/dev/raid/speed_limit_min
 }
